requirements.txt
Flask==3.0.3
PyJWT==2.10.1
requests==2.32.4
Werkzeug==3.0.3